refactor: attendeesessionService ì코드 리펙토링

+++ src/main/java/org/example/meeting/application/AttendeeSessionService.java
@@ -1,12 +1,13 @@
 package org.example.meeting.application;
 
 import lombok.RequiredArgsConstructor;
-import org.example.meeting.exception.attendee.AttendeeSessionCreationException;
-import org.example.meeting.exception.attendee.AttendeeSessionNotFoundException;
+import org.example.global.exception.type.NotFoundException;
 import org.example.meeting.domain.AttendeeSession;
+import org.example.meeting.exception.attendee.AttendeeExceptionType;
 import org.example.meeting.repository.AttendeeSessionRepository;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+
 import java.util.Optional;
 
 @Service
@@ -17,32 +18,29 @@ public class AttendeeSessionService {
     private final AttendeeSessionRepository attendeeSessionRepository;
 
     public void save(AttendeeSession attendeeSession) {
-        try {
-            attendeeSessionRepository.save(AttendeeSession.builder()
-                    .attendeeId(attendeeSession.getAttendeeId())
-                    .externalUserId(attendeeSession.getExternalUserId())
-                    .joinToken(attendeeSession.getJoinToken())
-                    .meetingId(attendeeSession.getMeetingId())
-                    .build());
-        } catch (Exception e) {
-            throw new AttendeeSessionCreationException("Error saving attendee session");
-        }
+        attendeeSessionRepository.save(AttendeeSession.builder()
+                .attendeeId(attendeeSession.getAttendeeId())
+                .externalUserId(attendeeSession.getExternalUserId())
+                .joinToken(attendeeSession.getJoinToken())
+                .meetingId(attendeeSession.getMeetingId())
+                .build());
     }
 
     public Optional<AttendeeSession> findByExternalUserId(String externalUserId) {
-        try {
             return attendeeSessionRepository.findByExternalUserId(externalUserId);
-        } catch (Exception e) {
-            throw new AttendeeSessionNotFoundException("Error finding attendee session by external user ID: " + externalUserId);
-        }
     }
 
     public void deleteByMeetingId(String meetingId) {
+        if (!attendeeSessionRepository.existsByMeetingId(meetingId)) {
+            throw new NotFoundException(AttendeeExceptionType.NOT_FOUND_ATTENDEE);
+        }
         attendeeSessionRepository.deleteByMeetingId(meetingId);
     }
 
     public void deleteByExternalUserId(String externalUserId) {
+        if (!attendeeSessionRepository.existsByExternalUserId(externalUserId)) {
+            throw new NotFoundException(AttendeeExceptionType.NOT_FOUND_ATTENDEE);
+        }
         attendeeSessionRepository.deleteByExternalUserId(externalUserId);
     }
-
 }
